refactor: renamed folder name refactor: method with validation
