+++ src/main/java/ru/daniil/split/config/AppService.java
@@ -1,24 +1,22 @@
-package ru.daniil.sawing.config;
+package ru.daniil.split.config;
 
-import ru.daniil.sawing.myExceptions.NonValidNameException;
-import ru.daniil.sawing.myExceptions.PersonIsExistException;
-import ru.daniil.sawing.myExceptions.ZeroSpendsException;
+import ru.daniil.split.exceptions.NonValidNameException;
+import ru.daniil.split.exceptions.PersonIsExistException;
+import ru.daniil.split.exceptions.ZeroSpendsException;
 
 import java.math.BigDecimal;
 import java.util.*;
 
 public class AppService {
 
-    public static final int MAX_NAME_LENGTH = 15;
-    public static final int MIN_NAME_LENGTH = 4;
+    private final int MAX_NAME_LENGTH = 15;
+    private final int MIN_NAME_LENGTH = 4;
     private int allSpends = 0;
     private final Set<String> allPersons = new HashSet<>();
 
     public void addNewPerson(String personName) throws NonValidNameException, PersonIsExistException {
         isValidPersonName(personName);
-        if (allPersons.contains(personName)) {
-            throw new PersonIsExistException("Person + {" + personName + "} already exist");
-        }
+        personIsExist(personName);
         allPersons.add(personName);
     }
 
@@ -42,19 +40,25 @@ public class AppService {
         return new BigDecimal(allSpends / allPersons.size());
     }
 
-    private static void isValidSpend(int newSpend) throws ZeroSpendsException {
+    private void isValidSpend(int newSpend) throws ZeroSpendsException {
         if (newSpend <= 0) {
             throw new ZeroSpendsException("Spending can't be negative");
         }
     }
 
-    private static void isValidPersonName(String personName) throws NonValidNameException {
+    private void isValidPersonName(String personName) throws NonValidNameException {
         if (personName.length() < MIN_NAME_LENGTH || personName.length() > MAX_NAME_LENGTH) {
             throw new NonValidNameException("The name cannot be shorter than 4 characters and longer than 15");
         }
         int wordCount = personName.split(" ").length;
         if (wordCount > 1) {
             throw new NonValidNameException("The person name can only consist of one word");
+        }
+    }
+
+    private void personIsExist(String personName) throws PersonIsExistException {
+        if (allPersons.contains(personName)) {
+            throw new PersonIsExistException("Person + {" + personName + "} already exist");
         }
     }
 }
