+++ src/main/java/ru/daniil/split/exceptions/PersonIsExistException.java
@@ -1,4 +1,4 @@
-package ru.daniil.sawing.myExceptions;
+package ru.daniil.split.exceptions;
 
 public class PersonIsExistException extends Exception {
     public PersonIsExistException(String s) {
